+++ src/App.tsx
@@ -10,9 +10,6 @@ const App: FC = () => (
   <BrowserRouter>
     <div>
       <Route exact path="/" component={RecruitmentIndexPage} />
-      <Route path="/card">
-        <RecruitmentCard title="募集タイトル" companyName="会社名" />
-      </Route>
       <Route path="/sign_in" component={SignInPage} />
     </div>
   </BrowserRouter>
