+++ src/components/Carousel/style.module.scss
@@ -1,0 +1,4 @@
+.carousel {
+  width: 100vw;
+  // margin-left: -27vw;
+}
