+++ src/scenes/recruitment/RecruitmentIndexPage/index.tsx
@@ -2,13 +2,12 @@ import React, {useState} from "react";
 import styles from "./style.module.scss";
 // import { Link } from "react-router-dom";
 import RecruitmentCard from "../../../components/RecruitmentCard";
-import Slider from "react-slick";
-import "slick-carousel/slick/slick.css";
-import "slick-carousel/slick/slick-theme.css";
+import Carousel from "../../../components/Carousel";
+
 import { isTemplateExpression } from "typescript";
+import { useCurrentAccount } from "../../../data/hooks/useCurrentAccount";
 
 const RecruitmentIndexPage = () => {
-  const [isLoggedIn, setIsLoggedIn] = useState(false);
   const [recommendations, setRecommnedations] = useState([
     {title: "おすすめ1", companyName: "会社1"},
     {title: "おすすめ2", companyName: "会社2"},
@@ -25,7 +24,11 @@ const RecruitmentIndexPage = () => {
     {title: "ぼしゅう4", companyName: "会社4"}
   ]);
 
-  const cardsList = (list: Array<{title: string, companyName: string}>) => (
+  const {account} = useCurrentAccount();
+
+  type RecruitmentCardProps = React.ComponentProps<typeof RecruitmentCard>;
+
+  const cardsList = (list: Array<RecruitmentCardProps>) => (
     list.map((item, index) => (
       <li key={index}>
         <div className={styles.wrapper}>
@@ -38,25 +41,15 @@ const RecruitmentIndexPage = () => {
     ))
   );
 
-  const settings = {
-    centerMode: true,
-    variableWidth: true,
-    speed: 500,
-    infinite: true,
-    slidesToShow: 1,
-    slidesToScroll: 1,
-    dots: false
-  };
-
   return (
     <>
-      {!isLoggedIn &&
+      {account &&
         <div className={styles.recommendations}>
           <h1>おすすめの募集</h1>
           <ul>
-              <Slider className={styles.slider} {...settings}>
-                {cardsList(recommendations)}
-              </Slider>
+            <Carousel
+              children={cardsList(recommendations)}
+            />
           </ul>
         </div>
       }
